+++ crates/effigy-manifest/src/bundles.rs
@@ -595,7 +595,7 @@ extensions = [
 catalog = "nginx"
 document_root = "."
 rewrite_all_to = "/vendor/genesis.php"
-asset_fallback = ""
+asset_fallback = "/vendor/genesis.php"
 error_page_404 = "/vendor/genesis.php"
 
 [containers.__CONTAINER_NAME__.services.db]
@@ -1065,6 +1065,9 @@ extensions = [
 catalog = "nginx"
 document_root = "."
 service = "{{ inputs.workspace_service_name }}"
+rewrite_all_to = "/vendor/genesis.php"
+asset_fallback = "/vendor/genesis.php"
+error_page_404 = "/vendor/genesis.php"
 
 [containers.{{ inputs.container_name }}.services.db]
 catalog = "mariadb"